refactor: Update Onboarding screen UI and improve previews
