+++ app/src/main/java/com/amsterdam/cutetudee/presentation/component/onboarding/OnboardingIndicators.kt
@@ -1,9 +1,8 @@
-package com.amsterdam.cutetudee.presentation.component
+package com.amsterdam.cutetudee.presentation.component.onboarding
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
-import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
@@ -12,24 +11,24 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import com.amsterdam.cutetudee.presentation.theme.AppTheme
 import com.amsterdam.cutetudee.presentation.theme.CuteTudeeTheme
-
+import com.amsterdam.cutetudee.presentation.utils.ThemeAndLocalePreviews
 
 @Composable
 fun OnboardingIndicators(
     modifier: Modifier = Modifier,
-    currentIndicator: Int
+    currentIndicator: Int,
+    numberOfIndicators: Int = 3
 ) {
     Row(modifier = modifier, horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-        (0..2).forEach { index ->
+        (0 until numberOfIndicators).forEach { index ->
             val background by animateColorAsState(
                 targetValue = if (currentIndicator == index) AppTheme.color.primary
                 else AppTheme.color.primaryVariant, tween(1000)
             )
-            Indicator(
+            Box(
                 modifier = Modifier
                     .height(5.dp)
                     .weight(1f)
@@ -41,15 +40,10 @@ fun OnboardingIndicators(
     }
 }
 
-@Composable
-private fun Indicator(modifier: Modifier = Modifier) {
-    Box(modifier.height(5.dp)) {}
-}
-
-@PreviewLightDark()
+@ThemeAndLocalePreviews ()
 @Composable
 private fun OnboardingCardPreviewDark() {
-    CuteTudeeTheme(isDarkTheme = isSystemInDarkTheme()) {
-         OnboardingIndicators(currentIndicator = 2)
+    CuteTudeeTheme() {
+        OnboardingIndicators(currentIndicator = 2)
     }
 }
